[2604-CHORE-001] promote VitalSign type and isVitalRecorded to lib/vitals.ts; fix LOS recorded predicate; normalize admin GET field name
